@@ -12,28 +12,27 @@
     "event router"
   ],
   "scripts": {
-    "test-ci": "jest --config jest.config.json --verbose",
+    "test-ci": "jest --verbose",
     "build-dev": "export NODE_ENV=dev&& webpack --display-reasons --display-error-details",
     "build-prod": "export NODE_ENV=prod&& webpack --display-reasons --display-error-details",
     "start": "webpack --watch",
-    "test": "jest --watchAll --no-cache --config jest.config.json"
+    "test": "jest --watchAll --no-cache"
   },
   "main": "dist/esp.js",
   "author": "Keith Woods",
   "license": "Apache-2.0",
   "repository": "https://github.com/esp/esp-js",
-  "typings": "./esp.d.ts",
+  "typings": "./.dist/typings/index.d.ts",
   "devDependencies": {
-    "@types/jest": "^23.0.0",
-    "awesome-typescript-loader": "5.0.0",
-    "babel-core": "^6.4.5",
-    "babel-jest": "23.0.1",
-    "babel-loader": "7.1.4",
+    "@types/jest": "23.1.6",
+    "babel-core": "^6.26.3",
+    "babel-jest": "^23.4.0",
+    "babel-loader": "7.1.5",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
     "babel-plugin-transform-flow-strip-types": "^6.5.0",
     "babel-plugin-transform-runtime": "^6.4.3",
     "babel-polyfill": "^6.20.0",
-    "babel-preset-es2015": "^6.3.13",
+    "babel-preset-node6": "^11.0.0",
     "babel-preset-react": "^6.3.13",
     "babel-preset-stage-0": "^6.3.13",
     "clean-webpack-plugin": "^0.1.19",
@@ -41,14 +40,37 @@
     "eslint-loader": "2.0.0",
     "inject-loader": "4.0.1",
     "jasmine-core": "3.1.0",
-    "jest": "^23.1.0",
+    "jest": "23.4.1",
+    "regenerator-runtime": "^0.12.0",
     "source-map-loader": "0.2.3",
+    "ts-loader": "4.4.2",
     "tslint": "^5.9.1",
     "tslint-loader": "^3.6.0",
-    "typescript": "^2.1.4",
-    "webpack": "4.12.0",
-    "webpack-cli": "^3.0.3",
+    "typescript": "2.9.2",
+    "webpack": "4.16.0",
+    "webpack-cli": "^3.0.8",
     "webpack-dev-server": "3.1.4"
   },
-  "dependencies": {}
+  "dependencies": {},
+  "jest": {
+    "globals": {
+      "ts-jest": {
+        "tsConfigFile": "tsconfig.json"
+      }
+    },
+    "setupFiles": [
+      "<rootDir>/tests/__jest__/configureEspLogging.ts"
+    ],
+    "transform": {
+      "^.+\\.jsx?$": "babel-jest",
+      "^.+\\.tsx?$": "<rootDir>/tests/__jest__/typeScriptPreprocessor.ts"
+    },
+    "testMatch": [
+      "**/tests/**/*Tests.[jt]s"
+    ],
+    "moduleFileExtensions": [
+      "ts",
+      "js"
+    ]
+  }
 }
