@@ -1,6 +1,6 @@
 import {Observer} from './Observer';
 import {Disposable} from '../system/disposables/disposable';
 
-export interface Subscribe {
-    (observer: Observer): Disposable;
+export interface Subscribe<T> {
+    (observer: Observer<T>): Disposable;
 }
