@@ -16,8 +16,8 @@
  */
  // notice_end
 
-export class ObservationStage {
-    static get preview() { return  'preview'; }
-    static get normal() { return  'normal'; }
-    static get committed() { return  'committed'; }
+export enum ObservationStage {
+    preview = 'preview',
+    normal = 'normal',
+    committed = 'committed'
 }
