@@ -18,90 +18,93 @@
 
 import {Router}  from './Router';
 import {Guard} from '../system';
-import {Observable} from '../reactive';
+import {Observable, RouterObservable} from '../reactive';
+import {ObservationStage} from './ObservationStage';
+import {EventEnvelope} from './envelopes';
+import {OnObserve} from '../reactive/Observable';
 
-export class SingleModelRouter {
+export class SingleModelRouter<TModel> {
     private _underlying: Router;
     private _targetModelId: string;
 
     public constructor() {
     }
 
-    public static create() {
-        let router = new SingleModelRouter();
+    public static create<TModel>() {
+        let router = new SingleModelRouter<TModel>();
         router._underlying = new Router();
         router._targetModelId = 'modelId';
         return router;
     }
 
-    public static createWithModel(model) {
+    public static createWithModel<TModel>(model: TModel) {
         Guard.isDefined(model, 'Model passed to to createWithModel must not be undefined.');
-        let router = new SingleModelRouter();
+        let router = new SingleModelRouter<TModel>();
         router._underlying = new Router();
         router._targetModelId = 'modelId';
         router.setModel(model);
         return router;
     }
 
-    public static createWithRouter(underlyingRouter, modelId) {
+    public static createWithRouter<TModel>(underlyingRouter: Router, modelId: string) {
         Guard.isString(modelId, 'The modelId should be a string.');
         if (!(underlyingRouter instanceof Router)) {
             throw new Error('underlyingRouter must be of type Router.');
         }
 
-        let router = new SingleModelRouter();
+        let router = new SingleModelRouter<TModel>();
         router._underlying = underlyingRouter;
         router._targetModelId = modelId;
         return router;
     }
 
-    setModel(model) {
+    setModel(model: TModel) {
         Guard.isDefined(model, 'Model passed to setModel() must not be undefined.');
         this._underlying.addModel(this._targetModelId, model);
     }
 
-    publishEvent(eventType, event) {
+    publishEvent(eventType: string, event: any) {
         this._ensureModelIsSet();
         this._underlying.publishEvent(this._targetModelId, eventType, event);
     }
 
-    executeEvent(eventType, event) {
+    executeEvent(eventType: string, event: any) {
         this._ensureModelIsSet();
         this._underlying.executeEvent(eventType, event);
     }
 
-    runAction(action) {
+    runAction(action: () => void) {
         this._ensureModelIsSet();
         this._underlying.runAction(this._targetModelId, action);
     }
 
-    getEventObservable(eventType, stage) {
+    getEventObservable<TEvent>(eventType: string, stage?: ObservationStage): Observable<EventEnvelope<TEvent, TModel>> {
         this._ensureModelIsSet();
-        return this._underlying.getEventObservable(this._targetModelId, eventType, stage);
+        return this._underlying.getEventObservable<TEvent, TModel>(this._targetModelId, eventType, stage);
     }
 
-    getModelObservable() {
+    getModelObservable(): Observable<TModel> {
         this._ensureModelIsSet();
-        return this._underlying.getModelObservable(this._targetModelId);
+        return this._underlying.getModelObservable<TModel>(this._targetModelId);
     }
 
-    createObservable(observer) {
+    createObservable(observer: OnObserve<TModel>): RouterObservable<TModel> {
         return Observable
             .create(observer)
             .asRouterObservable(this._underlying)
             .subscribeOn(this._targetModelId);
     }
 
-    createSubject() {
-        return this._underlying.createSubject();
+    createSubject<T>() {
+        return this._underlying.createSubject<T>();
     }
 
-    observeEventsOn(object, methodPrefix = '_observe_') {
+    observeEventsOn(object: any, methodPrefix = '_observe_') {
         this._ensureModelIsSet();
         return this._underlying.observeEventsOn(this._targetModelId, object, methodPrefix);
     }
 
-    isOnDispatchLoop() {
+    isOnDispatchLoop(): boolean {
         this._ensureModelIsSet();
         return this._underlying.isOnDispatchLoopFor(this._targetModelId);
     }
