@@ -3,6 +3,6 @@ import {Level} from './level';
 export interface LogEvent {
     logger: string;
     level: Level;
-    message: string,
-    args: any[]
+    message: string;
+    args: any[];
 }
