@@ -32,11 +32,11 @@ describe('Router', () => {
             expect(() => {_router.addModel('foo', undefined); }).toThrow();
             expect(() => {_router.addModel('foo', {}, 'not a function'); }).toThrow();
             expect(() => {_router.addModel({ },{ }); }).toThrow(new Error('The modelId argument should be a string'));
-            expect(() => {_router.addModel("modelId", { },"notSomeOptions"); }).toThrow(new Error('The options argument should be an object'));
-            expect(() => {_router.addModel("modelId", { }, { preEventProcessor: {} }); }).toThrow(new Error('preEventProcessor on the options parameter is neither a function nor an object with a process() method'));
-            expect(() => {_router.addModel("modelId", { }, { preEventProcessor: "boo" }); }).toThrow(new Error('preEventProcessor on the options parameter is neither a function nor an object with a process() method'));
-            expect(() => {_router.addModel("modelId", { }, { postEventProcessor:{}}); }).toThrow(new Error('postEventProcessor on the options parameter is neither a function nor an object with a process() method'));
-            expect(() => {_router.addModel("modelId", { }, { postEventProcessor:"boo"}); }).toThrow(new Error('postEventProcessor on the options parameter is neither a function nor an object with a process() method'));
+            expect(() => {_router.addModel('modelId', { },'notSomeOptions'); }).toThrow(new Error('The options argument should be an object'));
+            expect(() => {_router.addModel('modelId', { }, { preEventProcessor: {} }); }).toThrow(new Error('preEventProcessor on the options parameter is neither a function nor an object with a process() method'));
+            expect(() => {_router.addModel('modelId', { }, { preEventProcessor: 'boo' }); }).toThrow('preEventProcessor on the options parameter is neither a function nor an object with a process() method');
+            expect(() => {_router.addModel('modelId', { }, { postEventProcessor:{}}); }).toThrow(new Error('postEventProcessor on the options parameter is neither a function nor an object with a process() method'));
+            expect(() => {_router.addModel('modelId', { }, { postEventProcessor:'boo'}); }).toThrow(new Error('postEventProcessor on the options parameter is neither a function nor an object with a process() method'));
         });
 
         it('should throw if model already registered', () => {
