@@ -65,14 +65,14 @@ describe('Router', () => {
 
         it('should deliver correct model and event to target event observers', () => {
             let receivedModel2, receivedEvent2, model1ReceivedEvent = false;
-            _router.getEventObservable(_model1.id, "fooEvent").subscribe(() => {
+            _router.getEventObservable(_model1.id, 'fooEvent').subscribe(() => {
                 model1ReceivedEvent = true;
             });
-            _router.getEventObservable(_model2.id, "fooEvent").subscribe((event, context, model) => {
+            _router.getEventObservable(_model2.id, 'fooEvent').subscribe(({event, context, model}) => {
                 receivedModel2 = model;
                 receivedEvent2 = event;
             });
-            _router.publishEvent(_model2.id, "fooEvent", 1);
+            _router.publishEvent(_model2.id, 'fooEvent', 1);
             expect(receivedModel2).toBeDefined();
             expect(receivedModel2).toBe(_model2);
             expect(receivedEvent2).toBeDefined();
@@ -83,8 +83,8 @@ describe('Router', () => {
 
         it('should dispatch updates for the child model only', () => {
             let model1UpdateCount = 0, model2UpdateCount = 0;
-            _router.getEventObservable(_model1.id, "fooEvent").subscribe((event, context, model) => { /* noop */});
-            _router.getEventObservable(_model2.id, "fooEvent").subscribe((event, context, model) => { /* noop */});
+            _router.getEventObservable(_model1.id, 'fooEvent').subscribe(({event, context, model}) => { /* noop */});
+            _router.getEventObservable(_model2.id, 'fooEvent').subscribe(({event, context, model}) => { /* noop */});
             _router.getModelObservable(_model1.id).subscribe(() => {
                 model1UpdateCount++;
             });
@@ -93,7 +93,7 @@ describe('Router', () => {
                 model2UpdateCount++;
             });
             expect(model2UpdateCount).toEqual(1);
-            _router.publishEvent(_model2.id, "fooEvent", 1);
+            _router.publishEvent(_model2.id, 'fooEvent', 1);
             expect(model1UpdateCount).toEqual(1);
             expect(model2UpdateCount).toEqual(2);
         });
@@ -103,18 +103,18 @@ describe('Router', () => {
 
             it('throws if you observe a modelChangedEvent using a string identifier', () => {
                 expect(() => {
-                    _router.getEventObservable(_model2.id, 'modelChangedEvent').subscribe((event, context, model) => {});
-                }).toThrow(new Error("You can not observe a modelChangedEvent using only the eventType string. You must pass an object identifying the modelId to monitor. E.g. replace the eventType param with: { eventType: 'modelChangedEvent', modelId: 'yourRelatedModelId' }"));
+                    _router.getEventObservable(_model2.id, 'modelChangedEvent').subscribe(({event, context, model}) => {});
+                }).toThrow(new Error('You can not observe a modelChangedEvent using only the eventType string. You must pass an object identifying the modelId to monitor. E.g. replace the eventType param with: { eventType: \'modelChangedEvent\', modelId: \'yourRelatedModelId\' }'));
             });
 
             it('can observe a modelChangedEvent when child\'s event workflow done', () => {
-                _router.getEventObservable(_model1.id, "fooEvent").subscribe((event, context, model) => { /* noop */});
-                _router.getEventObservable(_model2.id, { eventType: "modelChangedEvent", modelId: _model1.id }).subscribe((event, context, model) => {
+                _router.getEventObservable(_model1.id, 'fooEvent').subscribe(({event, context, model}) => { /* noop */});
+                _router.getEventObservable(_model2.id, { eventType: "modelChangedEvent", modelId: _model1.id }).subscribe(({event, context, model}) => {
                     _receivedModel = model;
                     _receivedEvent = event;
                     _workflowDone = _model1OptionsHelper.modelsSentForPostProcessing.length === 1;
                 });
-                _router.publishEvent(_model1.id, "fooEvent", 1);
+                _router.publishEvent(_model1.id, 'fooEvent', 1);
                 expect(_receivedModel).toBe(_model2);
                 expect(_receivedEvent).toBeDefined();
                 expect(_receivedEvent.model).toBeDefined();
@@ -127,30 +127,30 @@ describe('Router', () => {
                 // model 2 observes model 1
                 let _model2ReceivedItems = [],
                     _model3ReceivedItems = [];
-                _router.getEventObservable(_model1.id, "fooEvent").subscribe((event, context, model) => { /* noop */});
-                _router.getEventObservable(_model2.id, { eventType: "modelChangedEvent", modelId: _model1.id }).subscribe((event, context, model) => {
+                _router.getEventObservable(_model1.id, 'fooEvent').subscribe(({event, context, model}) => { /* noop */});
+                _router.getEventObservable(_model2.id, { eventType: "modelChangedEvent", modelId: _model1.id }).subscribe(({event, context, model}) => {
                     _model2ReceivedItems.push({event, context, model});
                 });
                 // model 3 observes model 1 and 2
-                _router.getEventObservable(_model3.id, { eventType: "modelChangedEvent", modelId: _model1.id }).subscribe((event, context, model) => {
+                _router.getEventObservable(_model3.id, { eventType: "modelChangedEvent", modelId: _model1.id }).subscribe(({event, context, model}) => {
                     _model3ReceivedItems.push({event, context, model});
                 });
-                _router.getEventObservable(_model3.id, { eventType: "modelChangedEvent", modelId: _model2.id }).subscribe((event, context, model) => {
+                _router.getEventObservable(_model3.id, { eventType: "modelChangedEvent", modelId: _model2.id }).subscribe(({event, context, model}) => {
                     _model3ReceivedItems.push({event, context, model});
                 });
-                _router.publishEvent(_model1.id, "fooEvent", 1);
+                _router.publishEvent(_model1.id, 'fooEvent', 1);
 
                 expect(_model2ReceivedItems.length).toBe(1);
                 expect(_model3ReceivedItems.length).toBe(2);
             });
 
-            xit('detects and errors when models are observing each other', () => {
-                _router.getEventObservable(_model1.id, "fooEvent").subscribe((event, context, model) => { /* noop */});
-                _router.getEventObservable(_model1.id, { eventType: "modelChangedEvent", modelId: _model2.id }).subscribe((event, context, model) => {
+            it.skip('detects and errors when models are observing each other', () => {
+                _router.getEventObservable(_model1.id, 'fooEvent').subscribe(({event, context, model}) => { /* noop */});
+                _router.getEventObservable(_model1.id, { eventType: "modelChangedEvent", modelId: _model2.id }).subscribe(({event, context, model}) => {
                 });
-                _router.getEventObservable(_model2.id, { eventType: "modelChangedEvent", modelId: _model1.id }).subscribe((event, context, model) => {
+                _router.getEventObservable(_model2.id, { eventType: "modelChangedEvent", modelId: _model1.id }).subscribe(({event, context, model}) => {
                 });
-                _router.publishEvent(_model1.id, "fooEvent", 1);
+                _router.publishEvent(_model1.id, 'fooEvent', 1);
             });
         });
     });
