@@ -2,14 +2,14 @@
 /*
  * Copyright 2015 Dev Shop Limited
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
+ * Licensed under the Apache License, Version 2.0 (the 'License');
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
  *   http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
+ * distributed under the License is distributed on an 'AS IS' BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
@@ -43,8 +43,8 @@ describe('Router', () => {
             _modelRouter = esp.SingleModelRouter.createWithModel(_model);
 
             _dispatchedModelNumbers = [];
-            _modelRouter.getEventObservable('fooEvent').subscribe((e, c, m) => {
-                m.aNumber = e;
+            _modelRouter.getEventObservable('fooEvent').subscribe(({event, context, model}) => {
+                model.aNumber = event;
             });
             _modelRouter.getModelObservable().subscribe(m => {
                 _dispatchedModelNumbers.push(m.aNumber);
@@ -57,12 +57,12 @@ describe('Router', () => {
         });
 
         it('should proxy executeEvent to correct model event processor', ()=> {
-            _modelRouter.getEventObservable('barEvent').subscribe((e, c, m) => {
-                m.executePassed = m.anotherNumber === 0;
+            _modelRouter.getEventObservable('barEvent').subscribe(({event, context, model}) => {
+                model.executePassed = model.anotherNumber === 0;
             });
-            _modelRouter.getEventObservable('fooEvent2').subscribe((e, c, m) => {
+            _modelRouter.getEventObservable('fooEvent2').subscribe(({event, context, model}) => {
                 _modelRouter.executeEvent('barEvent', 'theBar');
-                m.anotherNumber = 1;
+                model.anotherNumber = 1;
             });
             _modelRouter.publishEvent('fooEvent2', {});
             expect(_model.executePassed).toEqual(true);
@@ -82,7 +82,7 @@ describe('Router', () => {
         it('should proxy isOnDispatchLoop', ()=> {
             _modelRouter.observeEventsOn(_model);
             let actualIsOnDispatchResult = null;
-            _modelRouter.getEventObservable('fooEvent').subscribe((e, c, m) => {
+            _modelRouter.getEventObservable('fooEvent').subscribe(({event, context, model}) => {
                 actualIsOnDispatchResult = _modelRouter.isOnDispatchLoop();
             });
             _modelRouter.publishEvent('fooEvent', {});
@@ -97,13 +97,13 @@ describe('Router', () => {
                 it('should throw if publishEvent used without setting model', () => {
                     let router = esp.SingleModelRouter.create();
                     expect(() => {
-                        router.publishEvent("foo", {});
+                        router.publishEvent('foo', {});
                     }).toThrow(_expectedError);
                 });
                 it('should throw if executeEvent used without setting model', () => {
                     let router = esp.SingleModelRouter.create();
                     expect(() => {
-                        router.executeEvent("foo", {});
+                        router.executeEvent('foo', {});
                     }).toThrow(_expectedError);
                 });
                 it('should throw if runAction used without setting model', () => {
